# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fill.me API
  description: |
    Fill.me provides a comprehensive REST API for form creation, management, and analytics. 
    This API enables developers to integrate form functionality into their applications and manage form data programmatically.
    
    ## Authentication
    
    Fill.me API supports multiple authentication methods:
    - **JWT Token**: `Authorization: Bearer YOUR_JWT_TOKEN`
    - **API Key**: `X-API-Key: YOUR_API_KEY`
    - **Basic Auth**: `Authorization: Basic base64(email:password)`
    
    ## Rate Limiting
    
    Rate limits vary by subscription plan:
    - Free: 100 requests/hour
    - Starter: 500 requests/hour
    - Professional: 2,000 requests/hour
    - Enterprise: 10,000 requests/hour
    
  version: 1.3.0
  contact:
    name: Fill.me Support
    email: support@fill.me
    url: https://fill.me/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.fill.me
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - BasicAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin, moderator]
          example: "user"
        subscription:
          type: object
          properties:
            plan:
              type: string
              enum: [free, starter, professional, enterprise]
            status:
              type: string
              enum: [active, cancelled, expired]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "The provided data is invalid"
        details:
          type: object
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Forms
...